SpriteBatch 구현 중

+++ Source/TGON/Graphics/Sprite.cpp
@@ -9,7 +9,7 @@ namespace tgon
 
 Sprite::Sprite(const std::shared_ptr<Texture>& texture) :
     m_texture(texture),
-    m_textureRect(0, 0, m_texture->GetWidth(), m_texture->GetHeight()),
+    m_textureRect(0, 0, static_cast<float>(m_texture->GetWidth()), static_cast<float>(m_texture->GetHeight())),
     m_enableScissorRect(false),
     m_layer(0),
     m_blendMode(BlendMode::Normal)
