+++ Source/TGON/Time/Time.cpp
@@ -16,7 +16,7 @@ TGON_API int64_t GetUnixTimestamp()
     clock_gettime(CLOCK_REALTIME, &ts);
 #endif
     
-    return (ts.tv_sec * 1000000000L + ts.tv_nsec) * 0.000001;
+    return static_cast<int64_t>((ts.tv_sec * 1000000000L + ts.tv_nsec) * 0.000001);
 }
 
 } /* namespace tgon */
